Intern comments in nullable type nodes
Summary:
Interns leading comments for nullable types. Comments that directly precede the `?` will be attached as leading comments. Comments between the `?` and type will be attached as leading comments for the type, and comments succeeding the type will be attached as trailing comments for the type:

```
/* L nullable */ ? /* L type */ string /* T type */
```

Note that this required creating a new record type in the AST for nullable type node so that we can keep track of both the type argument and the attached comments.

Reviewed By: mroch

Differential Revision: D20389800

fbshipit-source-id: 2a52d31a2cf187ca3b1628e09915fcbc7e883e85