Intern comments in array type nodes
Summary:
Interns trailing comments around array type nodes. Comments that directly succeed the `[]` will be attached as trailing comments. Comments before  the inner type or between the inner type and the brackets will be attached to the inner type:

```
/* L string */ string /* T string */ [] /* T array */
```

Note that this required creating a new record type in the AST for array type nodes so that we can keep track of both the inner types and the attached comments.

Reviewed By: gkz

Differential Revision: D20392949

fbshipit-source-id: a0b0a3a3653143b5464358d8eece55c6b2555dfc