Move tests/ to types-first
Summary:
In transitioning to types-first as the default mode for Flow, we are porting all existing
tests to use this mode.

Tests that have
```
classic_only: true
```
in their `.testconfig` could not be ported directly to types-first. `runtests.sh` will run
these tests in classic mode. There are currently 51 such tests (out of 641 tests for which
we run `flow check`, etc.). The reasons they cannot be immediately ported to types-first are:

* They exercise cases that become irrelevant in types-first:
  * MergedT (type_destructors_merged_t)
  * React.createClass (These ones are now too burdensome to annotate)
  * circular defs that become trivial in types-first, due to annotations (e.g. lint_suppressions_*)
  * tests on untyped exports (e.g. missing_annotation, type_visitor)
  * tests on modernizing code to types-first (autofix_insert_type)

* They exercise features that are buggy or TODOs in types-first (master task T64196901):
  * destructured exports (T55450302)
  * find-refs (T64150420)
  * computed properties (T64194787)
  * node_modules (T64195181)
  * saved-state (T64189656)
  * focusing dependencies (T64196619)
  * opaque type imports (T64197313)
  * module type imports (T64206661)
  * default export positioning (T64251568)

Reviewed By: samwgoldman

Differential Revision: D20454299

fbshipit-source-id: 9bf0a4a42efad5f150f47a31d621ede940e51498