Intern comments in union and intersection type nodes
Summary:
Interns leading comments in union and intersection type nodes. Union and intersection types consist of a list of types separated by either "`|`"s or "`&`"s. Since comments directly adjacent to the inner types will be attached to those inner types, the only leading comments that can appear on union/intersection nodes is if the union/intersection begins with a `|` or `&` character:

```
type Foo =
/* L union */
| /* L type */ Bar /* T type */
| /* L type */ Baz /* T type */
```

Union/intersection types that do not begin with a `|` or `&` will not have leading comments attached:

```
type Foo = /* L type */ Bar /* T type */ | /* L type */ Baz /* T type */
```

Note that this required creating a new node in the AST for unions and intersections, so that they can hold both the list of inner types as well as attached comments.

Reviewed By: mroch

Differential Revision: D20432727

fbshipit-source-id: 5005a6fc15a096d08acf4a933ffd1ae45079d2f4