Intern comments in function type nodes
Summary:
Interns leading comments around function type nodes when applicable.

Function type nodes may begin with a parenthesized params list, unparenthesized single param, or type params list. In the case of a parenthesized params list, leading comments will be attached to the params list:

```
/* L params */ (X) /* T params */ => Y
```

In the case of a single unparenthesized param, leading comments will be attached to that param's id:

```
/* L id */ X /* T id */ => Y
```

In the case of type params, leading comments will be attached to the type params. Note that comment attachment around type params has not yet been implemented, but will be implemented in a future change:
```
/* L tparams */ <T>(T) => Y
```

Trailing comments will always be attached to the return type. The full comment attachment for function types is shown below:

```
/* L tparams */ < /* L tparam */ T /* T tparam */> /* L params */ (/* L param */ param /* T param */) /* T params */ => /* L return */ return /* T return */
```

As described above, leading and trailing comments will usually be attached to inner nodes (the params/tparams and the return type). The only way for comments to be attached to the function type node itself is if the function type node is wrapped in parentheses. Support for comment attachment around parenthesized types will be implemented in an upcoming change, but for now let's add the `comments` field to the function type AST node to prepare for this.

This change also handles comments around callable properties in object types, as most of the code is shared with regular function types.

```
{ /* L params */ (/* L generic */ X /* T generic */) /* T params */ : /* L generic */ Y /* T generic */ };
```

Finally, the function type is also reused for method type annotations (both normal, getter/setter, and internal slot). None of these can have leading comments on the function type, as those leading comments would instead be attached to either the method name's identifier (or the internal slot).

```
{
  foo(X): Y
  get foo(X): Y
  [[foo]](X): Y
}
```

Reviewed By: mroch

Differential Revision: D20396676

fbshipit-source-id: 18584615828d955a12ef196f3bec5207f8df0910