Intern comments around generic types
Summary:
Interns leading and trailing comments around generic types. Generic types begin with an identifier, so leading comments will be attached to that inner identifier node (unless the entire generic is wrapped in parentheses). Generic types optionally end with a list of type parameters. If a generic type does not end with type parameters then trailing comments will be attached to the last identifier in the generic. If a generic does end with type parameters, trailing comments after the `>` will be attached to the generic node. Examples of comment attachment for generics:

```
/* L id */ Foo /* T id */
/* L id */ Foo /* T id */ </* L id */ T /* T id */> /* T generic */
/* L id */ Foo /* T id */ . /* L id */ Bar /* T id */
/* L generic */ (Foo) /* T generic */
```

This should virtually complete comment interning for type nodes.

Reviewed By: gkz

Differential Revision: D20451469

fbshipit-source-id: c2db310620ed888c2bdfaf998f518c3a70c916ff