Intern comments around parenthesized types
Summary:
If a type is wrapped in parentheses, comments directly leading or trailing those parentheses should be attached to the inner type (since there is no parenthesized type/expression node in the AST). This is accomplished the exact same way as for parenthesized expressions - after the inner type is parsed, if we detect that it was a parenthesized type we call the new `add_comments` helper which merges the comments around the parentheses with the comments inside the parentheses (if any exist).

```
/* L any */ (any) /* L any */
/ * L any */ (/* L any */ any /* T any */) /* T any */
```

The only type that does not yet handle comment attachment when surrounded by parentheses are generic types. This will be added in an upcoming change.

Reviewed By: mroch

Differential Revision: D20448442

fbshipit-source-id: f9023dc5bbb7375063a7f913c2e89c41ac0c967e