[autocomplete] remove param snippets
Summary:
when you autocomplete a function or method, we currently insert a snippet including the params.

this is not always desirable because functions are often passed around. you can sometimes tell by context, but not always. for consistency, it is more predictable if we never do it.

instead, signature help appears when you type the opening `(`, which provides the same information just without tab stops.

Fixes https://github.com/flowtype/flow-for-vscode/issues/363

Changelog:

- fix(autocomplete): Autocomplete no longer inserts snippets containing function arguments. These suggestions were sometimes unwanted, such as when passing around a function, and similar information is now provided by "signature help".

Reviewed By: gkz

Differential Revision: D20560353

fbshipit-source-id: 913175edbaa23970d36a6a93a46c127d94236141