[Likely to cause new errors] Fix a spread unsoundness in union and intersection spreads
Summary:
We have a special case in spreads to skip widening when we encounter an annotation. Widening is only necessary on values, since they may introduce constraint cycles that lead to exponential behavior, so avoiding widening whenever possible allows us to reliably produce more complete results without running into performance issues.

This special case had a bug when the annotation appeared in a union-- whenever we encountered an AnnotT ~> ObjKitT ObjectWiden constraint, we would immediately flow the AnnotT to the tout (which then triggers the spread) and we would **ignore the remaining unresolved tvars in the resolution tool in the ObjKitT**.

This diff restricts the special case to only when the annotation does not appear as part of a union or intersection.

We could make this more complete by flowing the annotation the tout _and_ continuing the resolution.

Reviewed By: samwgoldman

Differential Revision: D20375216

fbshipit-source-id: b93f6a6a1c79e17525e1e384f0bb42222fa6d63d