Store package.json errors in saved state
Summary:
D20285844 exposed an issue where no saved-state was getting generated when malformed `package.json` files existed. The `node_modules_with_bad_package_format` test began failing when the tests were run with saved state. Previously, it succeeded because even though there was no saved state file present, and the `--saved-state-no-fallback` flag was provided, the Flow server started up successfully anyway.

The cause of the failure was that there was an `Error ()` entry for the malformed `package.json` file. This caused the `save-state` command to fail with an exception. To fix this, we no longer assume that everything in the `PackageHeap` is `Ok` when generating saved-state, and instead store the full `result`.

Reviewed By: mroch

Differential Revision: D20501815

fbshipit-source-id: 656368f66eb8e9b777c8cacec156657c69db6580