Intern comments in export declarations
Summary:
Interns leading and trailing comments around export declarations, both named and default.

Both named and default exports can have leading comments, which are comments which directly precede the `export` keyword.

For named exports, if the export contains a declaration (variable, class, enum, function, interface, type alias, or opaque type alias) then trailing comments are attached to that inner declaration node. Named exports that do not contains a declaration (batch exported values/types or named exports with a list of specifiers) will have trailing comments attached to the export node itself, as there is no inner AST node that it makes sense to attach trailing comments to.

```
/* L export */ export default Foo; /* T export */
/* L export */ export default function () {} /* T func decl */
/* L export */ export {/* L id */ A /* T id */ as /* L id */ B /* T id */} from /* L lit */ 'D' /* T lit */; /* T export */
```

Note that we do not support comments between the `export` and `default` keywords, the `export` and `type` keywords (unless its a type alias), around the curly braces in named exports, or around the `*` token, as comments are not allowed in these locations by prettier.

Reviewed By: mroch

Differential Revision: D20495329

fbshipit-source-id: 6153109131aeb17ab09d4f366edc85974f95211d