Intern leading comments in functions
Summary:
Interns the leading comments around functions, up until the name of the function/params. Functions can have leading comments anywhere preceding/in between the `async`, `function`, and `*` tokens (depending on if the function is async or a generator).

```
/* L func */ async /* L func */ function /* L func */ * /* L id */ name /* T id */ (){}
```

Functions cannot normally have trailing comments, as if the function has a block body trailing comments will be attached as trailing that block, and if the function has an expression body trailing comments will be attached to that inner expression. However function expressions can have trailing comments if the entire function is wrapped in parentheses.

Note this change does not yet handle comment attachment around the params and type params, this will be added in an upcoming change.

The `Function` AST node is used in the following places, which are all affected by this change:
- Function expression
- Arrow functions
- Function declarations
- Class methods
- Object literal methods

Note that leading comments on getters and setters are not yet handled.

Reviewed By: mroch

Differential Revision: D20503460

fbshipit-source-id: 69ab0873346c25795d73c0c29b6734e86175f8e4