[Openness] openness graph
Summary:
This adds a new graph to the context to track Tvar openness. Openness works the following way.

First, there is a construction phase. Starting with an unresolved graph, tvars are added to the graph and connected to other tvars. If a tvar starts out `Unknown`, we mark it as `Derived` for blaming purposes, while if a tvar starts as either `Known` to be `Closed` or `Open`, then it is marked as `Inherent`.

Once all the tvars are added and all the connections made, we can resolve the graph. Starting at each of the `Inherent Open` nodes, we set all their `Unknown` dependents to "open", and accumulate a set of blame for each `Derived Open` node so that we know where their opennesses come from. Once the propagation is complete, the following will be true:

1) No `Open` node will ever have any `Unknown` dependents.
2) No `Unknown` node will have any `Open` dependencies.

This means that any remaining `Unknown` nodes can be set to `Closed`.

Reviewed By: mvitousek

Differential Revision: D20476203

fbshipit-source-id: 81ecb496ba8747da40e1e411974348c1ad82fe37