Add flag to LookupT for whether to give a suggestion upon failure
Summary:
Now that "did you mean?" suggestions in error messages are being presented to the user as quickfixes, it's important to limit these suggestions to only appear in places where there actually is a reasonable corresponding quickfix. This diff changes `LookupT`s' `ids` field to be an `option` so that we can specify whether or not we should emit a suggestion upon failure of this lookup.

In this diff all `LookupT`s are set to have `ids = Some <whatever it was before>` to align with the old behavior of always suggesting a fix when possible. In the next diff I use this to turn off suggestions in cases where the suggestions didn't make sense and caused bugs (object-to-object flows).

Reviewed By: dsainati1

Differential Revision: D20661718

fbshipit-source-id: 0a47453a999cff402fc451c0cd494d29a1de7f9d