Initial comment eating API and fix trailing comment attachment on statements
Summary:
We currently have an issue where trailing comments on one statement as also attached as leading comments of the next statement. This change introduces an initial comment eating API and uses it to begin a fix for this issue.

Comments that appear on the same line as the statement should be attached as trailing comments, wheras comments that appear on later lines but before the next statement should be attached as leading comments of the next statement:

```
/* L expr1 */ expr1; /* T expr1 stmt */
/* L expr2 */ expr2; /* T expr2 stmt */
```

If two statements appear on the same line, then the first statement will not have any trailing comments, as trailing comments will instead be attached as leading the second statement:

```
/* L expr */ expr; /* L expr */ expr; /* T expr stmt */
```

The difficulty comes about when there is an "implicit" semicolon that ends a statement. When the "implicit" semicolon is an `EOF` or an `R_CURLY` then we simply attach all trailing comments to the statement regardless of the lines they are on.

```
{
  expr
  /* T expr stmt */
}

expr
/* T expr stmt */
<EOF>
```

However there can also be an "implicit" semicolon if there is a line break after the end of the statement. Unfortunately, this poses a challenge since statements may end with an expression/type/etc. and have trailing comments on the next line greedily attached to that expression/type/etc. To fix this we must remove the trailing comments on that expression/type/etc that should actually be attached as leading the next statement. This change does not handle these more complicated cases of "implicit" semicolons - this will be fixed in an upcoming change.

## Parser API Changes

The main change to the parser API is removing `Eat.semicolon` (which was only used by the statement parser) in favor of a new `semicolon` helper function in the statement parser.

The new `Eat.comments_until_next_line` function takes in a line number and consumes all comments on the current token that start before or on the given line. This calls the new `Lookahead.junk_comments` function to remove (and return) all comments on the currently peeked-at token that start before the given line. This `Eat.comments_until_next_line` function is currently only used by the semicolon helper function for ending statements, but in the future will also be used to fix similar issues with ambiguous comment attachment in JSX and classes.

Reviewed By: mroch

Differential Revision: D20637954

fbshipit-source-id: dc68c3b54aae2438e40b6cd0ad338074d012e83d