[parser] Update usage of ast-types from 0.8 to 0.13, update custom ast types
Summary:
The version of `ast-types` we were using was very out of date. I've updated it, fixing a variety of usages that come from updating, and also removing all uneeded definitions from `custom_ast_types.js`. This defines what we need to send as a PR to `ast-types`, which I will do this coming week. (We would like to update `ast-types` because it is used as the base for JSCodeshift through `recast`).

Doing this identified several areas our parser diverged from ESTree and Babel Parser, which I have fixed lower down in the stack.

Reviewed By: mroch

Differential Revision: D20731500

fbshipit-source-id: 794e9747aaaed78153bde7975fdd1b4d4b4a1e4a