Introduce trailing comment remover and fix simple trailing comment attachment for statements
Summary:
We currently have an issue where trailing comments on one statement are also attached as leading comments of the next statement. This is made more difficult as some statements are closed by an explicit semicolon, others by an "implicit semicolon" (`EOF`, an `R_CURLY`, or a line break), and others by syntax specific to that statement.

When a statement ends in an explicit semicolon trailing comment attachment is straightforward - we simply need to attach comments after the semicolon as trailing the entire statement node.

Comment attachment is harder when a statement ends with an expression, type, or pattern node. This is because we greedily attach all comments after an expression, type, or pattern node to that node during parsing. This can attach comments that appear on later lines and should actually be attached to the next statement if this expression/type/pattern node ends its current statement. This can occur because the statement's optional semicolon was not supplied (e.g. an expression statement ended by a new line instead of an explicit semicolon), or if a statement usually ends with another AST node (e.g. a declare class node which ends in an object type).

For example, consider the case of attachment for an expression statement that ends with a line break instead of an explicit semicolon:

```
foo() /* L call */
/* T id */ bar()
```

We will greedily attach both comments after the call to the call node during parsing. We must then remove the trailing comments on the next line from the call node once we know this expression is the end of a statement.

This change introduces the "trailing comment remover", which is an AST mapper that removes all comments after a specified position. When we detect that a certain AST node that would have had trailing comments greedily attached is actually the end of a statement, we can pass it through the trailing comment remover to remove all comments that should not be trailing that node.

The trailing comments remover only needs to check the trailing comments, so we can short circuit key nodes to make it more efficient (e.g. block nodes only need to check their own trailing comments and don't need to recurse to child statements). I've implemented these optimizations for block and object type nodes so far, as these are the only non-trivial nodes needed in this change. Soon the remover will have to operate on all expressions and types (due to expression statements or type aliases not terminated by a semicolon), so future shortcut method implementations will be added in the future.

This change also introduces the `statement_end_trailing_comments` utility which is used to determine comment attachment when a statement ends with something other than a semicolon or other AST node (e.g. for switch statements or blocks, which both end with an `R_CURLY`). This utility returns both the correct trailing comments and a remover.

This change fixes trailing comment attachment follow the above rules for a number of statements where no complicated logic is needed.

## Break/Continue

Both break and continue statements may end with an optional semicolon, and may also include an optional label. If an explicit semicolon is used then comments after the semicolon will be attached as trailing the break/continue. If an explicit semicolon is not used and no label is supplied, comments after the break/continue will also be attached as trailing. If an explicit semicolon is not used and a label is present we must modify the trailing comments of the label id to have the correct trailing comments.

```
break /* T break */
break; /* T break */
break label /* T label */
break label; /* T break */
```

## Debugger

Debugger statements can have trailing comments directly after the `debugger` keyword, and also after the optional semicolon.

```
debugger /* T debug */
debugger /* T debug */ ; /* T debug */
```

## Do While

Do while statements have an optional closing semicolon:

```
do {} while (test) /* T do while */
do {} while (test) /* T do while */ ; /* T do while */
```

## Interface Declaration

Interface declarations end in an object type with no semicolon. This means we must unwrap the object type to attach the correct trailing comments.

## Declare Class

Declare class statements also end in an object type with no semicolon, so the same strategy used for interface declarations is used.

## Switch

Switch statements have a body that ends in an `R_CURLY` with no semicolon. This means we simply need to attach the correct trailing comments when parsing the closing `R_CURLY`.

## Try

Try statements always end in a block body for either a catch or finalizer. This means we simply need to detect which case we are in, and then attach the correct trailing comments to the block.

```
try {} catch {} /* T block */
try {} finally {} /* T block */
```

## Block

Blocks always end with an `R_CURLY`, so we simply need to attach the correct trailing comments when consuming the `R_CURLY`.

## Declare Module

Similarly, declare modules also always end with an `R_CURLY` so the same strategy as for blocks is employed.

## Import Declaration

Import declarations end in an optional semicolon, which is always preceded by a source string literal. If an explicit semicolon is present, comments after that semicolon will be attached as trailing the import declaration. If an explicit semicolon is not present then trailing comments will have been greedily attached to the source string literal, so we must unwrap it and add the correct trailing comments.

```
import "source" /* T string lit */
import "source"; /* T import */
import Foo from "source" /* T string lit */
import Foo from "source"; /* T import */
import {Foo} from "source" /* T string lit */
import {Foo} from "source"; /* T import */
```

Reviewed By: mroch

Differential Revision: D20653630

fbshipit-source-id: 74f3f17c37ce8707c1450580dda794ff126bc66b