Fix trailing comment attachent for statements that end in expressions
Summary:
This fixes trailing comment attachment with and without a semicolon for expression statements, return statements, and throw statements. All three of these statements end with an expression that is optionally followed by a semicolon. All trailing comments will greedily be attached the inner expression, however if the statement does not end with a semicolon this may accidentally attach comments on later lines that should be attached as leading the next statement. To fix this we can use the recently introduced comment remover to consume and remove trailing comments that should be attached to the next statement.

This change also fixes comment attachment on returns so that comments between the return and semicolon are not dropped, and are instead attached as trailing the return:

```
/* L return */ return /* T return */ ; /* T return */
```

This also exposed two bugs in the AST mapper - comments were not being properly handled for literals or tagged templates.

Since only a few expression nodes can contain trailing comments for that expression, we can greatly cut down on the number of unecessary calls in the remove comments mapper. Methods were overridden for the following nodes in the AST mapper, recursing into only the nodes that could potentially contain trailing comments.

- Array, Call, Import, JSX, New, Object, Type Cast: don't need to recurse into any children as comments will always be attached to this node itself
- Binary, Assignment, Logical: only need to recurse into right child
- Conditional: only need to recurse into alternate
- Functions, arrow functions: only have trailing comments on function node itself or body, do not need to recurse into other children (though trailing comments on function block bodies have not yet been fixed)
- Class: Can only have trailing comments on class node or class body node (though comments on class bodies have not yet been implemented)
- Member: only need to recurse into the property
- Sequence: only need to recurse into the very last expression in the sequence. Uses the new `id_list_last` helper function which maps a function over just the last element of a list, returning the original list if unchanged, and otherwise only modifying the last element.

Reviewed By: mroch

Differential Revision: D20709087

fbshipit-source-id: d1fe056caf1a27f0eb6b689626216a045792f30c