Intern comments around empty statements
Summary:
Interns leading and trailing comments around empty statements. Empty statements consist of a single semicolon, so leading comments are the coments preceding that semicolon and trailing comments are comments on the same line that succeed the semicolon (or on later lines if it is the last statement in its block/file).

This required making a new `Empty.t` record type that contains attached comments, as empty nodes did not previously have any fields.

Reviewed By: mroch

Differential Revision: D20655524

fbshipit-source-id: e6f80a7956fa57eb62c2f2397b69ad7294f36a76