Fix trailing comment attachment for type aliases with implicit semicolon
Summary:
Fixes trailing comment attachment for type alias statements that end in an implicit semicolon.

During parsing all trailing comments will greedily be attached the type node, however if the type alias does not end with a semicolon this may accidentally attach comments on later lines that should be attached as leading the next statement. To fix this we can use the recently introduced comment remover to consume and remove trailing comments that should be attached to the next statement.

Since only a few type nodes can contain trailing comments for that type, we can greatly cut down on the number of unecessary calls in the remove comments mapper. Methods were overridden for the following nodes in the AST mapper, recursing into only the nodes that could potentially contain trailing comments:

- Array Type, Tuple Type: trailing comments must always be attached to this node, so don't need to recurse into any children
- Function Type: can only have trailing comments on node itself or return type
- Interface Type: can only have trailing comments on node itself or object type body
- Generic Type: can only have trailing comments on the node itself, or the rightmost id. This means if the generic is qualified we do not need to recurse into the qualification.
- Union Type, Intersection Type: can only have trailing comments on node itself or the last type in the union/intersection. The types in union or intersection nodes are represnted as a tuple of the first type, the second type, and then a list of all remaining types. This means we need to check whether the list of remaining types is empty in which case we recurse into the second type. Otherwise, we recurse into the last element of the list of types using the `id_list_last` helper function to only modify the list if the last element is changed.

Reviewed By: mroch

Differential Revision: D20713339

fbshipit-source-id: ad3641e94974d7039ef9e682398c06de17868118