Fix trailing comment attachment for declare function statements
Summary:
Fixes trailing comment attachment for declare function statements with implicit semicolons.

If a declare function statement ends with an implicit semicolon, the last AST node in the statement will be either the return type or a predicate. During parsing all trailing comments will be greedly attached to this last inner AST node. However, this may accidentally attach comments on later lines that should be attached as leading the next statement. To fix this we can use the recently introduced comment remover to consume and remove trailing comments that should be attached to the next statement.

Reviewed By: mroch

Differential Revision: D20717561

fbshipit-source-id: b54552d0fc5a4e6fb7086924952e335b86a75217