Fix trailing comment attachment for variable declaration statements
Summary:
Fixes trailing comment attachment for variable declaration statements with implicit semicolons.

During parsing all expression and type nodes will have all trailing comments greedily attached. This poses a problem for statements that end in an expression or type node, as comments will be attached to these inner expression/type nodes even if they appear on a later line and should actually be attached to the next statement. To fix this we can use the recently introduced comment remover to consume and remove trailing comments that should be attached to the next statement.

When variable declarations do not end with an explicit semicolon, comments will be attached to the last declarator. The final node in this declarator will be an expression if an initializer is supplied. If not, the final node will be a type if a type annotation is supplied. If not, the final node will be the identifier for the variable name.

```
const x = 1; /* T var decl */
const x = 1 /* T num */
const x: any /* T any */
const x /* T id */
```

This change introduces a `variable_declaration_end` helper that to encapsulate logic shared across var, const, and let declarations.

We can also optimize the comments remover by having it only recurse into the id/type node if there is no init node.

Reviewed By: mroch

Differential Revision: D20717999

fbshipit-source-id: a1bd74421332fda95417eac6a8d969a9debd95ec