Fix trailing comment attachment on function bodies
Summary:
We currently greedily attach all comments after the closing `R_CURLY` of a function block body to that block. However, if this is a function declaration then this may attach comments on later lines that should instead be attached to the next statement.

The correct attachment here is to only attach all comments after the closing `R_CURLY` if the function is an expression. For declarations/methods/etc. only trailing comments on the same line as the `R_CURLY` should be eaten and attached - comments on later lines should be attached to the next statement (unless the next token is an `R_CURLY` or `EOF` signaling the end of a block).

Reviewed By: mroch

Differential Revision: D20743024

fbshipit-source-id: 8e0d362ed677f6083cbe39c4cdcbeb5384b71612