Distribute $Exact over unions
Summary:
$Exact<T|U> should be treated as $Exact<T>|$Exact<U>, on both the left and the
right. The behavior is unchanged when appearing on the left, except that we
might find more union optimizations. On the right, this is a bug fix.

Before:
T -> $Exact<A|B>
A|B -> MakeExact (Lower T)
A -> MakeExact (Lower T)
T -> $Exact<A>
B -> MakeExact (Lower T)
T -> $Exact<B>

which is actually more like T -> $Exact<A>&$Exact<B>, which is not the correct
behavior.

After:
T -> $Exact<A|B>
A|B -> MakeExact (Lower T)
T -> $Exact<A>|$Exact<B>
... union speculation ...

Reviewed By: dsainati1

Differential Revision: D20841984

fbshipit-source-id: 155352ab9a0a661c13e0b9dbc3e46d47b4602907