Fix duplicate comment attachment within JSX tags
Summary:
We currently greedily attach leading and trailing comments when parsing, however this leads to comments being attached to two nodes when a comment is ambiguously between two AST nodes. We must specify rules to disambiguate which node these comments should be attached to.

This change fixes duplicate comment attachment within JSX, which was happening due to comments between adjacent "JSX tag component" nodes (JSX tag names, attributes, and spread attributes). The general rule we are following is that if the "JSX tag component" is the last component before the tag ends (i.e. it is followed by an `T_DIV` or  `T_GREATER_THAN`), then all trailing comments should be attached. If instead the component is followed by another JSX tag component on another line then all trailing comments on the first component's line will be attached as trailing the first component, and all comments on later lines will be attached as leading the next component. Finally if a component is followed by another component on the same line, all comments will be attached as leading the next component.

As an example, comments are currently attached multiple times like this:

```
<div /* T name, L attribute */
/* T name, L attribute */ name=value /* T attribute, L spread */
/* T attribute, L spread */ {...x} /* T spread, L attribute */
/* T spread, L attribute */ name
/>
```

And when following the new rules, comments should actually be attached like this:

```
<div /* T name */
/* L attribute */ name=value /* T attribute */
/* L spread */ {...x} /* T spread */
/* L attribute */ name
/>
```

This change introduces the `tag_component_trailing_comments` helper which implements the rules described above. The only other change is making sure this function is called when we are at the end of a tag component.

For spread attributes this is very easy to tell, as after we have parsed the closing `R_CURLY` we know we are at the end of the spread attribute.

Regular attributes with values are also very easy, as we know we are at the end after we have parsed either the `T_JSX_TEXT` for a string literal value or the closing `T_RCURLY` of an expression container value.

The tag components that are a bit more difficult to detect the end of are those that end in identifiers, namely the tag name and shorthand attributes without values. Luckily we can detect if an identifier is the end of a tag component by checking the next token. If it is a `T_COLON` or `T_PERIOD` this identifier is the first part of a namespaced name or member expression, respectively (e.g. `First:Second` or `First.Second`), so it is not the end of a tag component and all trailing comments should be attached. Similarly if the identifier is followed by a `T_ASSIGN` then it is the end of an attribute name (with value), so it is not the end of a tag component and all trailing comments should be attached.

If the identifier is followed by another token then it must be either the end of the initial tag name, or the end of an attribute without a value.

Reviewed By: mroch

Differential Revision: D20747493

fbshipit-source-id: 6204adfc0b96fa3bc0756e11c0e3d4545ec51859