Lookup $Exports<> types directly in builtins
Summary:
This diff changes the resolution logic for $Exports<> types. Before, we would
look up the module in the environment, which falls back to a builtins lookup.
Instead, this changes the logic to look up in the builtins directly.

Note that this is a subtle breaking change. Before, you could declare a module
outside of the libdefs and have a $Exports<> type resolve to that module in the
same file.

I think it's fair to break this, since declaring a module outside of libdefs is
not well-supported overall. For example, you also can not `require()` such a
module.

Reviewed By: panagosg7

Differential Revision: D17425492

fbshipit-source-id: 3ec2229b0391d073ca3de2f812ce789265cc38c7