Intern comments in class methods
Summary:
Attaches comments to class method nodes in the AST.

Methods have a number of optional keywords that they can start with - `static`, `async`, and `*` (for generators). Comments before any of these keywords will be attached as leading the entire method.

```
/* L meth */ static /* L meth */ async /* L meth */ * /* L id */ method() {}
```

Getters and setters will also have comments preceding the `get` or `set` keyword attached as leading the method.

```
/* L meth */ get /* L id */ method() {}
/* L meth */ set /* L id */ method(x) {}
```

The rest of comment attachment for methods is shown below. Just like regular functions comments will be attached tightly to the method name and method params, and trailing comments will be attached to the block body.

```
/* L id */ method /* T id */ <T> /* L params */ () /* T params */ {} /* T block */
```

Also note that estree will not output computed keys in method names, so to avoid dropping comments in the estree output let's merge comments attached to the computed key with the comments on the method definition node itself.

Reviewed By: mroch

Differential Revision: D20779142

fbshipit-source-id: 1bc490362c00373cc5518417133a25e5fc74a03e